requirements.txt
appnope==0.1.0
boto==2.42.0
bz2file==0.98
click==6.6
cycler==0.10.0
decorator==4.0.10
docopt==0.6.2
entrypoints==0.2.2
Flask==0.11.1
gensim==0.13.2
ipykernel==4.4.1
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
itsdangerous==0.24
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.3.0
jupyter-console==5.0.0
jupyter-core==4.1.1
langdetect==1.0.6
MarkupSafe==0.23
matplotlib==1.5.2
mistune==0.7.3
mysql-connector-python==2.0.4
mysql-connector-repackaged==0.3.1
nbconvert==4.2.0
nbformat==4.1.0
nltk==3.2.1
notebook==4.2.2
numpy==1.11.1
pandas==0.18.1
patsy==0.4.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.3.1
prompt-toolkit==1.0.7
ptyprocess==0.5.1
Pygments==2.1.3
pyparsing==2.1.8
python-dateutil==2.5.3
pytz==2016.6.1
pyzmq==15.4.0
qtconsole==4.2.1
requests==2.11.1
scikit-learn==0.17.1
scipy==0.18.0
seaborn==0.7.1
simplegeneric==0.8.1
six==1.10.0
smart-open==1.3.4
statsmodels==0.6.1
terminado==0.6
tornado==4.4.1
traitlets==4.2.2
Unidecode==0.4.19
wcwidth==0.1.7
Werkzeug==0.11.10
widgetsnbextension==1.2.6
wordcloud==1.2.1
xgboost==0.6a2
